+++ semana16/aula46/create-view-template-aula47.sql
@@ -1,0 +1,85 @@
+--## Exercício 1-A
+--raw, que permite que enviemos uma
+--query para o banco usando a linguagem SQL diretamente.
+
+--1-b
+/*
+const searchActor = async (name: string): Promise<any> => {
+  const result = await connection.raw(`
+    SELECT * FROM Actor WHERE name = "${name}"
+  `)
+  return result
+}
+*/
+
+--1-c
+/*
+const Actors = async (gender: string): Promise<any> => {
+  const result = await connection.raw(`
+    SELECT COUNT(*) as count FROM Actor WHERE gender = "${gender}"
+  `);
+	
+  const count = result[0][0].count;
+  return count;
+};
+*/
+--## Exercício 2-A
+/*
+const updateActor = async (id: string, salary: number): Promise<any> => {
+  await connection("Actor")
+    .update({
+      salary: salary,
+    })
+    .where("id", id);
+};
+*/
+--2-b
+/*
+const deleteActor = async (id: string): Promise<void> => {
+  await connection("Actor")
+    .delete()
+    .where("id", id);
+}; 
+*/
+--2-c
+/*
+const avgSalary = async (gender: string): Promise<any> => {
+  const result = await connection("Actor")
+    .avg("salary as average")
+    .where({ gender });
+
+  return result[0].average;
+};
+*/
+--## Exercício 3-A
+
+--3-b
+--3-c
+
+--## Exercício 4-A
+/*
+app.put("/actor", async (req: Request, res: Response) => {
+  try {
+    await updateSalary(req.body.id, req.body.salary);
+    res.status(200).send({
+      message: "Success",
+    });
+  } catch (err) {
+    res.status(400).send({
+      message: err.message,
+    });
+  }
+});
+*/
+--4-b
+/*
+app.delete("/actor/:id", async (req: Request, res: Response) => {
+  try {
+    await deleteActor(req.params.id);
+  } catch (err) {
+    res.status(400).send({
+      message: err.message,
+    });
+  }
+});
+*/
